QV-29 final code check... Data protection and Accessibility

+++ enseignant/delete_categories.php
@@ -1,6 +1,11 @@
 <?php
+    session_start();
+    if (!isset($_SESSION['user_id'])) {
+        header("Location: ../auth/login.php");
+        exit();
+    }
     require_once "../config/database.php";
-
+    
     if(isset($_POST['deleteCategory'])){
         $category_id = $_POST['category_id'];
         $stmt = $conn->prepare("DELETE FROM category WHERE id = ? ");
